+++ compilador.py
@@ -1,90 +1,6 @@
 from argparse import ArgumentParser
-import re
-
-#Classes
-class Token:
-    def __init__(self, funcao, id):
-        self.funcao = funcao
-        self.id = id
-
-class Simbolo:
-    def __init__(self,id,posicao,linha,coluna):
-        self.id = id
-        self.posicao = posicao
-        self.linha = linha
-        self.coluna = coluna
-
-    def __str__(self):
-        return "[{}, {}, ({}, {})]".format(self.id, "" if self.posicao==None else self.posicao, self.linha, self.coluna)
-
-class Dicionario:
-    def __init__(self):
-        self.lista = []
-        self.lista.append(Token("rem",61))
-        self.lista.append(Token("input",62))
-        self.lista.append(Token("let",63))
-        self.lista.append(Token("print",64))
-        self.lista.append(Token("goto",65))
-        self.lista.append(Token("if",66))
-        self.lista.append(Token("end",67))
-        self.lista.append(Token("=",11))
-        self.lista.append(Token("+",21))
-        self.lista.append(Token("-",22))
-        self.lista.append(Token("*",23))
-        self.lista.append(Token("/",24))
-        self.lista.append(Token("%",25))
-        self.lista.append(Token("==",31))
-        self.lista.append(Token("!=",32))
-        self.lista.append(Token(">",33))
-        self.lista.append(Token("<",34))
-        self.lista.append(Token(">=",35))
-        self.lista.append(Token("<=",36))
-
-        self.lista.append(Token("LF",10))
-        self.lista.append(Token("ETX",3))
-        self.lista.append(Token("var",41))
-        self.lista.append(Token("const",51))
-
-    def token(self,txt):
-        for item in self.lista:
-            if item.funcao==txt or (txt.isnumeric() and int(txt)==item.id):
-                return item
-        return Token("",0)
-
-    def token_existe(self,txt):
-        for item in self.lista:
-            if item.funcao==txt or (txt.isnumeric() and int(txt)==item.id):
-                return True
-        return False
-    
-
-#Funções úteis..
-def remover_espacos_duplos(txt):
-    while "  " in txt:
-        txt = re.sub('[ ]{2,}'," ",txt)
-    
-    return txt
-
-def prepara_linha(linha):
-    if linha=="":
-        return ""
-
-    linha = remover_espacos_duplos(linha)
-
-    if linha[0]==" ":
-        linha = linha[1:]
-    if linha[-1]==" ":        
-        linha = linha[0:-1]
-
-    return linha + " "
-
-def adiciona_se_nao_existe(var,txt):
-    if txt in var:
-        return var.index(txt)
-    else:
-        var.append(txt)
-        return len(var)-1
-    return -1
+import classes
+import functions
 
 #Lê o nome o arquivo e coloca na variavel args.filename
 parser = ArgumentParser()
@@ -92,7 +8,7 @@ parser.add_argument("-f", "--file", dest="filename", help="compile FILE", metava
 args = parser.parse_args()
 
 #Início do compilador
-dicionario = Dicionario()
+dicionario = classes.Dicionario()
 memoria = []
 resultado = []
 erros = []
@@ -107,7 +23,7 @@ for linha in range(1,len(linhas)+1):
     if conteudo=="":
         continue
 
-    conteudo = prepara_linha(conteudo)
+    conteudo = functions.prepara_linha(conteudo)
     buffer = ""
     comentario = False
 
@@ -133,28 +49,28 @@ for linha in range(1,len(linhas)+1):
 
                         numero_linha = int(buffer)
                         
-                        indice = adiciona_se_nao_existe(memoria, buffer)
-                        resultado.append(Simbolo(51, indice, linha, coluna-len(buffer)))
+                        indice = functions.adiciona_se_nao_existe(memoria, buffer)
+                        resultado.append(classes.Simbolo(51, indice, linha, coluna-len(buffer)))
                 else:
                     t = dicionario.token(buffer)
 
                     if t.id==0:
                         if buffer.isnumeric():
-                            resultado.append(Simbolo(51, None, linha, coluna-len(buffer)))
+                            resultado.append(classes.Simbolo(51, None, linha, coluna-len(buffer)))
                         else:
-                            indice = adiciona_se_nao_existe(memoria, buffer)
-                            resultado.append(Simbolo(41, indice, linha, coluna-len(buffer)))
+                            indice = functions.adiciona_se_nao_existe(memoria, buffer)
+                            resultado.append(classes.Simbolo(41, indice, linha, coluna-len(buffer)))
                     else:
-                        resultado.append(Simbolo(t.id, None, linha, coluna-len(buffer)))
+                        resultado.append(classes.Simbolo(t.id, None, linha, coluna-len(buffer)))
                         if t.id==61:
                             comentario=True
                 buffer = ""
 
         if coluna==len(conteudo):
             if linha==len(linhas):
-                resultado.append(Simbolo(3, None, linha, coluna))
+                resultado.append(classes.Simbolo(3, None, linha, coluna))
             else:
-                resultado.append(Simbolo(10, None, linha, coluna))
+                resultado.append(classes.Simbolo(10, None, linha, coluna))
 
 #print(erros)
 for simbolo in resultado:
